+++ common/remote/bitbucket/bitbucket_test.go
@@ -2,8 +2,8 @@ package bitbucket
 
 import (
 	"net/http"
+	"net/url"
 	"os"
-	"path/filepath"
 	"testing"
 
 	"github.com/golang/protobuf/jsonpb"
@@ -40,45 +40,22 @@ func TestBitbucket_FindWebhooksEmpty(t *testing.T) {
 	}
 }
 
-func TestBBTranslate_TranslatePR(t *testing.T) {
-	pwd, err := os.Getwd()
+func TestBitbucket_PostPRComment(t *testing.T) {
+	vcsConf := &pb.VCSCreds{
+		ClientId: "VEhMhdw6uprevzh8Du",
+		ClientSecret: "JtxmvZy3QR2dJQwbnwmVktCJa4jaVsJS",
+		TokenURL: "https://bitbucket.org/site/oauth2/access_token",
+		AcctName: "level11consulting",
+		SubType: pb.SubCredType_BITBUCKET,
+		Identifier: pb.SubCredType_BITBUCKET.String() + "_level11consulting",
+	}
+	client, _, err := GetBitbucketClient(vcsConf)
 	if err != nil {
 		t.Error(err)
-		return
 	}
-	file, err := os.Open(filepath.Join(pwd, "test-fixtures", "1528734656-pr-bb.json"))
-	defer file.Close()
-	trans := GetTranslator()
-	pull, err := trans.TranslatePR(file)
+	err = client.PostPRComment("level11consulting/go-til", "2", "1234", pb.StageResultVal_FAIL, 1234)
 	if err != nil {
 		t.Error(err)
-		return
-	}
-	if pull.Source.Hash != "dc128f78cd34" {
-		t.Error(test.StrFormatErrors("source hash", "dc128f78cd34", pull.Source.Hash))
-	}
-	if pull.Source.Branch != "newbranch" {
-		t.Error(test.StrFormatErrors("source branch", "newbranch", pull.Source.Branch))
-	}
-	if pull.Id != 1 {
-		t.Error(test.GenericStrFormatErrors("pr id", 1, pull.Id))
-	}
-	commits := "https://api.bitbucket.org/2.0/repositories/jessishank/mytestocy/pullrequests/1/commits"
-	if pull.Urls.Commits != commits {
-		t.Error(test.StrFormatErrors("commits url", commits, pull.Urls.Commits))
-	}
-	comments := "https://api.bitbucket.org/2.0/repositories/jessishank/mytestocy/pullrequests/1/comments"
-	if pull.Urls.Comments != comments {
-		t.Error(test.StrFormatErrors("comments url", comments, pull.Urls.Comments))
-	}
-	if pull.Title != "Newbranch" {
-		t.Error(test.StrFormatErrors("titile", "Newbranch", pull.Title))
-	}
-	if pull.Destination.Branch != "master" {
-		t.Error(test.StrFormatErrors("dest branch", "master", pull.Destination.Branch))
-	}
-	if pull.Destination.Hash != "32ed49560d10" {
-		t.Error(test.StrFormatErrors("dest hash", "32ed49560d10", pull.Destination.Hash))
 	}
 }
 
@@ -107,5 +84,9 @@ func (mhc MockHttpClient) PostUrl(url string, body string, unmarshalObj proto.Me
 }
 
 func (mhc MockHttpClient) GetUrlResponse(url string) (*http.Response, error) {
+	return nil, nil
+}
+
+func (mhc MockHttpClient)  PostUrlForm(url string, form url.Values) (*http.Response, error) {
 	return nil, nil
 }
