+++ models/bitbucket/common.proto
@@ -19,4 +19,7 @@ message Links {
     LinkUrl commits = 4;
     LinkUrl comments = 5;
     LinkUrl statuses = 6;
+    LinkUrl approve = 7;
+    LinkUrl decline = 8;
+    LinkUrl merge = 9;
 }
