+++ models/build-protos.sh
@@ -35,6 +35,7 @@ echo "injecting custom tags"
 protoc-go-inject-tag -input=models/pb/guideocelot.pb.go
 protoc-go-inject-tag -input=models/pb/build.pb.go
 protoc-go-inject-tag -input=models/pb/creds.pb.go
+protoc-go-inject-tag -input=models/bitbucket/pb/commonevententities.pb.go
 
 # then we have to run go get in the stub directory cause grpc gateway ¯\_(ツ)_/¯ does this even work
 cd models/pb
