+++ models/vcshandler.go
@@ -1,9 +1,10 @@
 package models
 
 import (
-	"io"
 	"fmt"
+	"io"
 
+	ocenet "github.com/shankj3/go-til/net"
 	pb "github.com/shankj3/ocelot/models/pb"
 	// ugh stuck 4 now
 	pbb "github.com/shankj3/ocelot/models/bitbucket/pb"
@@ -56,6 +57,11 @@ type VCSHandler interface {
 	//   unmarshal into its vcs-specific model, then translate to the global model to return a list of generic commits
 	GetPRCommits(url string) ([]*pb.Commit, error)
 
+	// PostPRComment will add a comment to a pr belonging to acct/repo acctRepo and id prId with a comment that is along the lines of
+	// Ocelot build has <status>.
+	PostPRComment(acctRepo, prId, hash string, failed bool, buildId int64) error
+
+	GetClient() ocenet.HttpClient
 }
 
 type Translator interface {
